requirements.txt
apt-clone==0.2.1
apturl==0.5.2
bcrypt==3.1.7
blinker==1.6.2
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.6
colorama==0.4.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
Flask==2.3.2
Flask-Cors==4.0.0
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jetson.GPIO==2.1.1
Jinja2==3.1.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.3
matplotlib==3.1.2
monotonic==1.5
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
PAM==0.4.2
pandas==0.25.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pyserial==3.5
python-apt==2.0.1
python-dateutil==2.7.3
python-dbusmock==0.19
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.3.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
systemd-python==234
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
urllib3==1.25.8
urwid==2.0.1
wadllib==1.3.3
Werkzeug==2.3.6
xkit==0.0.0
zipp==3.16.2
